requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.13.0
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via -r requirements.in
click==8.1.7
    # via uvicorn
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via pydantic
exceptiongroup==1.2.0
    # via anyio
fastapi==0.104.1
    # via -r requirements.in
greenlet==3.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
loguru==0.7.2
    # via -r requirements.in
mako==1.3.0
    # via alembic
markupsafe==2.1.3
    # via mako
pydantic[email]==2.5.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.14.5
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pyenchant==3.2.2
    # via python-camelcaser
python-camelcaser==1.0.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via pydantic-settings
sniffio==1.3.0
    # via anyio
sqlalchemy[asyncio]==2.0.23
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
typing-extensions==4.9.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.24.0.post1
    # via -r requirements.in
